New nexthop-module
The nexthop-module is responsible to the nexthop-handling (aka mail-forwarding
to a configurable destination).

The nexthop-module is controlled by the nexthop-configuration-switch. If a pathname
is configured, then the module writes the mail into the file (which can also be a
pipe or similar). Otherwise the module assumes that a hostname (with an optinal
port-number) is configured, and it delivers the mail to this host by SMTP.